Major restructure to the flow of UI in the app.
A major restructure to the organization of the UI in the hope that it will become easier to use. The main activities now are:
1. home (no particular functionality yet)
2. personalization - the player is allowed to choose a name, a color, and perhaps an icon. these will be useful especially in multi-player games.
3. code editor - more or less the same (added a confirm dialog when you select 'clear')
4. training (this is local playing, using your defined code)
5. online challenge - not implemented (but the idea is to allow to browse a list of online challenges and join a selected one.

This has also required the introduction of a new model class: Challenge. It is basically intended to model everything needed for a competition, including the grid and other limitations/features.