Android changes 2
Included selection of algorithm in Challenge JSON files. Algorithm cannot be changed, mazes will have to be regenerated to use different algorithms. Changed JSON values of rewards and penalties to Low Medium and High instead. Challenge class is now responsible to determine the amount of objects generated, according to the grid's size and the option selected. Maze editing and saving completed.